Update lock move indexer to own file and update test
