+++ packages/enclave-sdk/tests/indexer.rs
@@ -25,25 +25,19 @@ async fn test_indexer() -> Result<()> {
         .await?;
 
     let contract = Enclave::deploy(provider).await?;
-
-    println!("Contract addr: {}", contract.address());
     let address = contract.address().to_string();
     let endpoint = anvil.ws_endpoint();
-    let store = Arc::new(RwLock::new(InMemoryStore::new()));
-    // TODO: Builder might be appropriate here.
-    let mut indexer = EnclaveIndexer::new(&endpoint, &address, store.clone()).await?;
+
+    let mut indexer = EnclaveIndexer::new(&endpoint, &address, InMemoryStore::new()).await?;
+
     indexer.initialize().await?;
 
-    tokio::spawn(async move {
-        if let Err(er) = indexer.start().await {
-            eprintln!("Error running indexer: {}", er);
-        }
-    });
+    indexer.start()?;
 
+    // E3Activated
     let e3_id = 10;
     let expiration = 10;
     let public_key = vec![1, 2, 3, 4, 5, 6, 7, 8, 9];
-
     contract
         .emitE3Activated(
             Uint::from(e3_id),
@@ -57,10 +51,10 @@ async fn test_indexer() -> Result<()> {
 
     sleep(Duration::from_millis(1)).await;
 
+    // InputPublished
     let data = "Random data that wont actually be a string".to_string();
     let data_hash = 1234;
     let index = 1;
-
     contract
         .emitInputPublished(
             Uint::from(e3_id),
@@ -72,18 +66,12 @@ async fn test_indexer() -> Result<()> {
         .await?
         .watch()
         .await?;
+
     sleep(Duration::from_millis(1)).await;
 
-    let Some(e3) = store
-        .read()
-        .await
-        .get::<E3>(&format!("e3:{}", e3_id))
-        .await?
-    else {
-        panic!("Could not get e3");
-    };
+    let e3 = indexer.get_e3(e3_id).await?;
 
     assert_eq!(e3.ciphertext_inputs.len(), 1);
-    println!("{:?}", e3.ciphertext_inputs[0].0);
+
     Ok(())
 }
